Move index stringifier to hist manager
